Black hole now reduces gravity and pulls you in Added BPE meteors to black hole destruction Added the allowExplosions gamerule Added some black hole BRAAAAAAMMSS Players now float towards the black hole at the end Refactoring

+++ src/main/java/com/sp/destruction/DestructionEvent.java
@@ -25,8 +25,7 @@ public abstract class DestructionEvent {
         if(animation != null) {
             if (this.isActive()) {
                 this.progress++;
-//                this.progress = 2200;
-                animation.updateKeyframeAnimation((float) this.progress / this.duration);
+                animation.updateKeyframeAnimation((double) this.progress / this.duration);
 
             } else {
                 animation.resetAnimation();
